refactor: utils-> view, MoneyDTO -> view.MoneyDTO
