+++ src/main/java/view/UserInput.java
@@ -1,4 +1,4 @@
-package utils;
+package view;
 
 import java.util.*;
